pyproject.toml
[project]
name = "1-log-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "pyupgrade>=3.21.2",
]

[tool.isort]
profile = "black"
line_length = 176
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true

[tool.black]
line-length = 176

target-version = [
    'py312',
    'py313'
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

strict = true
